+++ android/src/main/java/de/atroo/securestorage/SecureStoragePlugin.java
@@ -1,0 +1,22 @@
+package de.atroo.securestorage;
+
+import com.getcapacitor.JSObject;
+import com.getcapacitor.Plugin;
+import com.getcapacitor.PluginCall;
+import com.getcapacitor.PluginMethod;
+import com.getcapacitor.annotation.CapacitorPlugin;
+
+@CapacitorPlugin(name = "SecureStoragePluginImpl")
+public class SecureStoragePlugin extends Plugin {
+
+    private SecureStoragePluginImpl implementation = new SecureStoragePluginImpl();
+
+    @PluginMethod
+    public void echo(PluginCall call) {
+        String value = call.getString("value");
+
+        JSObject ret = new JSObject();
+        ret.put("value", implementation.echo(value));
+        call.resolve(ret);
+    }
+}