+++ android/src/main/java/de/atroo/securestorage/SecureStoragePluginImpl.java
@@ -1,0 +1,11 @@
+package de.atroo.securestorage;
+
+import android.util.Log;
+
+public class SecureStoragePluginImpl {
+
+    public String echo(String value) {
+        Log.i("Echo", value);
+        return value;
+    }
+}