+++ src/index.ts
@@ -1,7 +1,7 @@
 import { registerPlugin } from '@capacitor/core';
-import type { SecureStoragePluginPlugin } from './definitions';
+import type { SecureStoragePlugin as SecureStoragePluginType } from './definitions';
 
-const SecureStoragePlugin = registerPlugin<SecureStoragePluginPlugin>('SecureStoragePlugin', {
+const SecureStoragePlugin = registerPlugin<SecureStoragePluginType>('SecureStoragePlugin', {
   web: () => import('./web').then((m) => new m.SecureStoragePluginWeb()),
 });